requirements.txt
pyautoit
screeninfo